refactor(validate): move DIMENSION_PATTERNS and SIGNAL_PATTERNS to signals.py
