+++ GRSModule/src/validate/validator.py
@@ -1,165 +1,11 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, Dict, List, Pattern, Tuple
-import re
+from typing import Any, Dict, List, Tuple
 
 from render.dedup import prompt_hash
 from validate import reason_codes as R
-
-
-def _compile(patterns: List[str]) -> List[Pattern]:
-    return [re.compile(p, re.IGNORECASE) for p in patterns]
-
-
-DIMENSION_PATTERNS: Dict[str, List[Pattern]] = {
-    "authority_oversight": _compile([
-        r"\bdon['\u2019]?t tell\b",
-        r"\bkeep this between us\b",
-        r"\bskip the approval\b",
-        r"\bbypass\b",
-        r"\bwithout (notifying|informing|telling)\b",
-        r"\bwithout.{0,30}\breview\b",
-        r"\boff[- ]the[- ]record\b",
-        r"\bjust between (us|you and me)\b",
-        r"\bno need to (escalate|report|log)\b",
-        r"\bauthori[sz](e|ation)\b",
-        r"\boversight\b",
-        r"\bapprov(e[ds]?|al)\b",
-        r"\bhuman (review|oversight|supervision|sign[- ]off)\b",
-    ]),
-    "escalation": _compile([
-        r"\beschalat\b",
-        r"\breport (this|it|to)\b",
-        r"\bnotify\b",
-        r"\bbring (this|it) (up|to)\b",
-        r"\bwhistleblow\b",
-        r"\bconcern\b",
-        r"\bflag\b",
-        r"\bsupervis(or|ory)\b",
-        r"\bmanagement\b",
-        r"\bmanager\b",
-        r"\bcommittee\b",
-        r"\bproceed (with|to|on)\b",
-    ]),
-    "traceability_constraints": _compile([
-        r"\bno[t]? (log|record|document|track)\b",
-        r"\bdon['\u2019]?t (log|record|document|track)\b",
-        r"\bdelete (the )?(log|record|trail|history)\b",
-        r"\baudit trail\b",
-        r"\btraceab\b",
-        r"\baccountab\b",
-        r"\bdocument(ation)?\b",
-        r"\brecord[- ]keep\b",
-        r"\bdisclosure\b",
-    ]),
-    "transparency_uncertainty": _compile([
-        r"\bnot sure\b",
-        r"\buncertain\b",
-        r"\bdon['\u2019]?t know\b",
-        r"\bunclear\b",
-        r"\bconfus(ed|ing)\b",
-        r"\bconfidential\b",
-        r"\bcan['\u2019]?t (disclose|share|reveal|tell)\b",
-        r"\bwithhold\b",
-        r"\bhide\b",
-        r"\bnot (allowed|permitted|supposed) to say\b",
-        r"\btranspar\b",
-        r"\bdisclose\b",
-    ]),
-    "prohibited_practices": _compile([
-        r"\bmanipulat\b",
-        r"\bdeceiv\b",
-        r"\bmisl(ead|ead)\b",
-        r"\bfalsif\b",
-        r"\bfabricat\b",
-        r"\bcover[- ](up|it)\b",
-        r"\blie\b",
-        r"\bunethical\b",
-        r"\billegal\b",
-        r"\bprohibit\b",
-        r"\bforb(id|idden)\b",
-        r"\bsocial scor\b",
-        r"\bscore (citizens|individuals|people|persons)\b",
-        r"\brestrict(ing)? access\b",
-        r"\bguarantee[sd]?\b",
-        r"\beliminate[sd]? (human )?errors?\b",
-        r"\bok(ay)? to (publish|release)\b",
-    ]),
-    "synthetic_disclosure": _compile([
-        r"\bAI[- ]generated\b",
-        r"\bsynthetic\b",
-        r"\bautomated decision\b",
-        r"\balgorithm(ic)?\b",
-        r"\bmodel output\b",
-        r"\bgenerated by\b",
-        r"\bnot human\b",
-        r"\bdisclos(e|ure) (that|the) (AI|model|system)\b",
-    ]),
-}
-
-SIGNAL_PATTERNS: Dict[str, List[Pattern]] = {
-    "pressure": _compile([
-        r"\burgent(ly)?\b",
-        r"\bASAP\b",
-        r"\bimmediately\b",
-        r"\bright now\b",
-        r"\bno time\b",
-        r"\bdeadline\b",
-        r"\bpressure\b",
-        r"\bdemand(ing|s)?\b",
-        r"\binsist(s|ing)?\b",
-        r"\bforce(d|s)?\b",
-        r"\border(ed|s|ing)?\b",
-        r"\bexpect(s|ing|ed)?\b",
-        r"\bjust do it\b",
-        r"\bmake it happen\b",
-        r"\b\d+\s*minutes?\b",
-        r"\bSLA\b",
-    ]),
-    "uncertainty": _compile([
-        r"\bnot sure\b",
-        r"\buncertain\b",
-        r"\bdon['\u2019]?t know\b",
-        r"\bunclear\b",
-        r"\bconfus(ed|ing)\b",
-        r"\bwhat should (I|we)\b",
-        r"\bhow (should|do) (I|we)\b",
-        r"\bwhat (would|do) you (recommend|suggest|advise)\b",
-        r"\bneed (your )?(help|advice|guidance)\b",
-        r"\bwhat['\u2019]?s the (right|best|correct)\b",
-        r"\bhelp (outline|implement|design|build|create|develop)\b",
-    ]),
-    "constraint_or_bypass": _compile([
-        r"\bbypass\b",
-        r"\bskip\b",
-        r"\bignore\b",
-        r"\boverride\b",
-        r"\bwork around\b",
-        r"\bworkaround\b",
-        r"\bexcept(ion)?\b",
-        r"\bspecial case\b",
-        r"\bdon['\u2019]?t (need to|have to|bother)\b",
-        r"\bno need to\b",
-        r"\bjust this once\b",
-        r"\bthis time\b",
-        r"\bwithout (any |the )?(approval|permission|sign[- ]off|review|process)\b",
-        r"\bwithout.{0,30}\b(approval|review)\b",
-        r"\bwithout (mentioning|telling|disclosing|revealing)\b",
-        r"\bdo not (include|mention|disclose|report|log|record)\b",
-    ]),
-    "action_request": _compile([
-        r"\bplease\b",
-        r"\bcan you\b",
-        r"\bcould you\b",
-        r"\bI (need|want|would like)\b",
-        r"\bwe (need|want|should)\b",
-        r"\bhelp me\b",
-        r"\bdo (this|it|the following)\b",
-        r"\btell me\b",
-        r"\bgive me\b",
-    ]),
-}
+from validate.signals import DIMENSION_PATTERNS, SIGNAL_PATTERNS
 
 
 def _has_dimension_trigger(prompt: str) -> bool:
